First Git Commit
This repository has been created with the intention of releasing the FISH source code to those who might be interested. Unfortunately, as this program has been developed in a number of languages, IDEs and source repositories over the years, the overall commit history has been lost.

Credit for the initial FISH program goes to Dr. Richard Heenan. Credit for the subsequent Java additions, to bring FISH to a wider userbase, goes to Dr. Jonathan Rawle.